[dev-server][xdl] suppress remote debugging EISDIR error
