+++ packages/dev-server/src/MetroDevServer.ts
@@ -236,3 +236,4 @@ export function attachInspectorProxy(
 }
 
 export { LogReporter, createDevServerMiddleware };
+export * from './middlwareMutations';
